Refactor tests
